+++ fiddles/vue/fiddle-0002-AgGridSparklineTs/src/main.ts
@@ -1,11 +1,12 @@
-import { createApp } from 'vue'
-import { createPinia } from 'pinia'
+import {createApp} from 'vue'
+import {createPinia} from 'pinia'
 
 import App from './App.vue'
 import router from './router'
 
 import './assets/main.css'
 import 'bootstrap/dist/css/bootstrap.css'
+import '@vueform/slider/themes/default.css'
 
 const app = createApp(App)